getExtension(NameReference)

+++ owsi-alfresco/owsi-alfresco-app/owsi-alfresco-app-core/src/main/java/fr/openwide/alfresco/app/core/node/binding/AbstractFileRepositoryContentSerializer.java
@@ -27,6 +27,9 @@ public abstract class AbstractFileRepositoryContentSerializer
 	@Override
 	public File deserialize(RepositoryNode node, NameReference contentProperty, InputStream inputStream) throws IOException {
 		File file = getFile(node);
+		if (file.exists()) {
+			throw new IllegalStateException("The file " + file.getAbsolutePath() + " already exists.");
+		}
 		try (OutputStream outputStream = new FileOutputStream(file)) {
 			IOUtils.copy(inputStream, outputStream);
 		}
